+++ src/services/api/sensitivityCalculator/SensCalcResultsMock.tsx
@@ -1,0 +1,106 @@
+export const SENSCALC_CONTINUUM_MOCKED = {
+  id: 1,
+  title: 'DUMMY',
+  section1: [
+    {
+      field: 'continuumSensitivityWeighted',
+      units: 'μJy/beam',
+      value: '193.39869461690563'
+    },
+    {
+      field: 'continuumConfusionNoise',
+      units: 'μJy/beam',
+      value: '1.5571964705842967'
+    },
+    {
+      field: 'continuumTotalSensitivity',
+      units: 'μJy/beam',
+      value: '193.4049635877299'
+    },
+    {
+      field: 'continuumSynthBeamSize',
+      units: 'arcsec²',
+      value: '5.50 x 2.86'
+    },
+    {
+      field: 'continuumSurfaceBrightnessSensitivity',
+      units: 'K',
+      value: '375.85865050198123'
+    }
+  ],
+  section2: [
+    {
+      field: 'spectralSensitivityWeighted',
+      units: 'mJy/beam',
+      value: '42.282426012622985'
+    },
+    {
+      field: 'spectralConfusionNoise',
+      units: 'μJy/beam',
+      value: '2.6504800547538223'
+    },
+    {
+      field: 'spectralTotalSensitivity',
+      units: 'mJy/beam',
+      value: '42.282426095695854'
+    },
+    {
+      field: 'spectralSynthBeamSize',
+      units: 'arcsec²',
+      value: '5.92 x 3.96'
+    },
+    {
+      field: 'spectralSurfaceBrightnessSensitivity',
+      units: 'K',
+      value: '55096.90764904839'
+    }
+  ],
+  section3: [
+    {
+      field: 'integrationTime',
+      units: 'h',
+      value: '1'
+    }
+  ],
+  statusGUI: 0
+};
+
+export const SENSCALC_SPECTRAL_MOCKED = {
+  id: 1,
+  title: 'DUMMY',
+  section1: [
+    {
+      field: 'spectralSensitivityWeighted',
+      units: 'mJy/beam',
+      value: '828.5057118237343'
+    },
+    {
+      field: 'spectralConfusionNoise',
+      units: 'μJy/beam',
+      value: '2.6504800547538223'
+    },
+    {
+      field: 'spectralTotalSensitivity',
+      units: 'mJy/beam',
+      value: '828.505711827974'
+    },
+    {
+      field: 'spectralSynthBeamSize',
+      units: 'arcsec²',
+      value: '5.92 x 3.96'
+    },
+    {
+      field: 'spectralSurfaceBrightnessSensitivity',
+      units: 'K',
+      value: '1079599.8930615226'
+    }
+  ],
+  section3: [
+    {
+      field: 'integrationTime',
+      units: 'h',
+      value: '1'
+    }
+  ],
+  statusGUI: 0
+};